+++ worker/src/lib.rs
@@ -7,7 +7,7 @@ pub async fn main(req: Request, env: Env) -> Result<Response> {
     Router::new()
         .get_async("/", |req, _ctx| async move {
             let (latitude, longitude) = req.cf().coordinates().unwrap();
-            let region = find_region_nearby(latitude as f64, longitude as f64);
+            let region = find_region_nearby(latitude, longitude);
             Response::ok(region)
         })
         .run(req, env)
